fixed whitespaces, added swagger types, fixed validation, fixed types in requests

+++ Looms.PoS/Looms.PoS.Application/Models/Responses/ServiceResponse.cs
@@ -7,7 +7,7 @@ public class ServiceResponse
     public string Category { get; init; } = string.Empty;
     public string Description { get; init; } = string.Empty;
     public decimal Price { get; init; }
-    public decimal DurationMin { get; init; }
+    public int DurationMin { get; init; }
     public Guid BusinessId { get; init; }
     public Guid TaxId { get; init; }
     public bool IsDeleted { get; init; } = false;
